feature(navsat_transform): Parameter to publish "utm" as parent frame
In order to create a transformation tree with multiple robots, a common
parent frame is required. So far, it was only possible to publish "utm"
as child frame of the odometry frame. Adding this feature, the user can
specify whether to publish the "utm" transform as parent or child of the
odometry frame.