feature(initialization): Initialization update for frames and parameters
In the constructor, the robot localization listener waits for incoming
message to initialize properly. This enables the exposition of the
getFrame id methods. Also added an additional parameter_namespace
parameter to refer to the robot localization instance so that the same
filter parameters can be used, this value defaults to ~ which implies
that we search in the local namespace.