Switch robot_localization to modern CMake idioms.
Some of these have been best practice for a while, and
some of them will become best practice in the near future.

1. Move the include directory down one-level.  This is to better
support overlays and underlays; see https://github.com/ros2/ros2/issues/1150
for the entire saga.  The upshot is that all includes should be
one more directory level down, and CMake will handle the differences.
This has been best practice since Humble.

2. Switch from ament_target_dependencies() to target_link_libraries().
ament_target_dependencies was developed in the days before
target_link_libraries() was fully capable, and nowadays
target_link_libraries() is a superset of the functionality.  There is
one oddity here, in that in order to deal with ROS message packages, we
have to use ${<msg_pkg_name>_TARGETS}, rather than the traditional CMake
target like msg_pkg_name::msg_pkg_name.  This is a bug in ROS that will
eventually be fixed.

3. Export a CMake target from robot_localization.  This means that
downstream packages will be able to use
target_link_libraries(<target> robot_localization::robot_localization)

4. Export all dependencies with ament_export_dependencies.  This ensures
that downstream project which rely on this one will be able to find
all of the includes.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>